fix: final setup

+++ scripts/run_miniworld.py
@@ -100,7 +100,7 @@ def train(  # noqa: PLR0912, PLR0915
     with initialize_config_dir(str(config.parent.absolute()), version_base=None):
         cfg = compose(config_name=config.name, overrides=list(overrides))
     cfg = Config.model_validate(cfg)
-    fabric = Fabric()
+    fabric = Fabric(precision="bf16-mixed")
     fabric.launch()
     if cfg.train.seed is not None:
         fabric.seed_everything(cfg.train.seed)
